+++ Cargo.toml
@@ -40,6 +40,8 @@ rand = "0.8.5"
 indexmap = { version = "2", features = ["serde"], default-features = true }
 arboard = "3.3.2"
 zeno = { version = "0.3.1", optional = true }
+syntect = "5.2.0"
+lazy_static = "1.5.0"
 
 [dependencies.uuid]
 version = "1.1.2"
