+++ src/theui/thestyle/classic.rs
@@ -46,7 +46,7 @@ impl TheStyle for TheClassicStyle {
     fn draw_text_edit_border(
         &mut self,
         buffer: &mut TheRGBABuffer,
-        widget: &mut dyn TheWidget,
+        widget: &dyn TheWidget,
         shrinker: &mut TheDimShrinker,
         ctx: &mut TheContext,
         embedded: bool,
